feat-cyh-test

+++ back-end/src/main/java/com/ssafy/back_end/main/controller/MainController.java
@@ -23,9 +23,9 @@ import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Tag (name = "메인페이지")
+@Tag(name = "메인페이지")
 @RestController
-@RequestMapping (value = "/api/main")
+@RequestMapping(value = "/api/main")
 public class MainController {
     private static final Logger log = LoggerFactory.getLogger(SnsItemController.class);
     private final MainService mainService;
@@ -37,10 +37,10 @@ public class MainController {
         this.workoutRoutineService = workoutRoutineService;
     }
 
-    @Operation (summary = "유저 정보 조회")
-    @GetMapping ("/profile/info")
+    @Operation(summary = "유저 정보 조회")
+    @GetMapping("/profile/info")
     public ResponseEntity<?> getUserInfo(HttpServletRequest request) {
-        int ID = (Integer)request.getAttribute("userId");
+        int ID = (Integer) request.getAttribute("userId");
         UserInfoDto userInfoDto = mainService.getUserInfo(ID);
 
         if (userInfoDto != null) {
@@ -49,11 +49,11 @@ public class MainController {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("유저 정보 조회 오류");
     }
 
-    @Operation (summary = "유저 프로필 수정")
-    @PutMapping ("/profile")
-    public ResponseEntity<?> modifyProfile(HttpServletRequest request, @RequestPart ("nickname") String nickname,
-                                           @RequestPart ("image") MultipartFile image) {
-        int ID = (Integer)request.getAttribute("userId");
+    @Operation(summary = "유저 프로필 수정")
+    @PutMapping("/profile")
+    public ResponseEntity<?> modifyProfile(HttpServletRequest request, @RequestPart("nickname") String nickname,
+                                           @RequestPart("image") MultipartFile image) {
+        int ID = (Integer) request.getAttribute("userId");
         String imageUrl = "";
 
         int result = mainService.checkNickname(nickname, ID);
@@ -70,8 +70,7 @@ public class MainController {
                 if (!isCreated) {
                     log.debug("[MainController] 디렉토리 {} 생성 실패", uploadDirectory.getAbsolutePath());
                     return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("디렉토리 생성 실패");
-                }
-                else {
+                } else {
                     log.debug("[MainController] 디렉토리 {} 생성 완료", uploadDirectory.getAbsolutePath());
                 }
             }
@@ -101,8 +100,7 @@ public class MainController {
 
                     log.debug("[MainController] 이미지 업로드 성공: {}", newFileName);
 
-                }
-                catch (IOException e) {
+                } catch (IOException e) {
                     log.error("[MainController] 이미지 업로드 오류", e);
                     return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("이미지 업로드 오류");
                 }
@@ -112,26 +110,27 @@ public class MainController {
             //등록 성공 시점
 
             String preImage = mainService.getImagePathsByUserId(ID);
+
+            if (preImage != null)
             // 2. 각 이미지 파일을 호스트 디렉토리에서 삭제
-            try {
-                File file = new File(preImage);
-                if (file.exists()) {
-                    if (file.delete()) {
-                        log.debug("이미지 파일 삭제 성공, {}", preImage);
+            {
+                try {
+                    File file = new File(preImage);
+                    if (file.exists()) {
+                        if (file.delete()) {
+                            log.debug("이미지 파일 삭제 성공, {}", preImage);
+                        } else {
+                            log.warn("이미지 파일 삭제 실패, {}", preImage);
+                            throw new RuntimeException("이미지 파일 삭제 실패: " + preImage);
+                        }
+                    } else {
+                        log.warn("이미지 파일이 존재하지 않습니다, {}", preImage);
+                        throw new RuntimeException("이미지 파일이 존재하지 않음: " + preImage);
                     }
-                    else {
-                        log.warn("이미지 파일 삭제 실패, {}", preImage);
-                        throw new RuntimeException("이미지 파일 삭제 실패: " + preImage);
-                    }
+                } catch (Exception e) {
+                    log.error("게시물 삭제 중 오류 발생: {}", e.getMessage());
+                    throw e;  // 트랜잭션 롤백을 위해 예외를 다시 던짐
                 }
-                else {
-                    log.warn("이미지 파일이 존재하지 않습니다, {}", preImage);
-                    throw new RuntimeException("이미지 파일이 존재하지 않음: " + preImage);
-                }
-            }
-            catch (Exception e) {
-                log.error("게시물 삭제 중 오류 발생: {}", e.getMessage());
-                throw e;  // 트랜잭션 롤백을 위해 예외를 다시 던짐
             }
 
             // 삭제 성공 시점
@@ -141,8 +140,7 @@ public class MainController {
                 return ResponseEntity.ok("유저 프로필 수정 완료");
             }
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("유저 프로필 수정 오류");
-        }
-        else {
+        } else {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("사용중인 닉네임입니다.");
         }
     }
@@ -172,10 +170,10 @@ public class MainController {
 //        }
 //    }
 
-    @Operation (summary = "회원정보 수정")
-    @PutMapping ("/profile/info")
+    @Operation(summary = "회원정보 수정")
+    @PutMapping("/profile/info")
     public ResponseEntity<?> modifyUserInfo(HttpServletRequest request, @RequestBody UserInfoDto userInfoDto) {
-        int ID = (Integer)request.getAttribute("userId");
+        int ID = (Integer) request.getAttribute("userId");
         userInfoDto.setId(ID);
         int result = mainService.modifyUserInfo(userInfoDto);
 
@@ -185,10 +183,10 @@ public class MainController {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("회원정보 수정 오류");
     }
 
-    @Operation (summary = "패스워드 변경")
-    @PutMapping ("/profile/password")
+    @Operation(summary = "패스워드 변경")
+    @PutMapping("/profile/password")
     public ResponseEntity<?> modifyPassword(HttpServletRequest request, @RequestBody UserPasswordDto userPasswordDto) {
-        int ID = (Integer)request.getAttribute("userId");
+        int ID = (Integer) request.getAttribute("userId");
         int result = mainService.modifyPassword(ID, userPasswordDto.getCurPassword(), userPasswordDto.getNewPassword());
 
         if (result != 0) {
@@ -197,10 +195,10 @@ public class MainController {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("패스워드 변경 오류");
     }
 
-    @Operation (summary = "회원탈퇴")
-    @DeleteMapping ("/profile")
+    @Operation(summary = "회원탈퇴")
+    @DeleteMapping("/profile")
     public ResponseEntity<?> deleteUser(HttpServletRequest request) {
-        int ID = (Integer)request.getAttribute("userId");
+        int ID = (Integer) request.getAttribute("userId");
         int result = mainService.deleteUser(ID);
 
         if (result != 0) {
@@ -209,8 +207,8 @@ public class MainController {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("회원탈퇴 오류");
     }
 
-    @Operation (summary = "모든 사용자의 루틴 목록 조회", description = "모든 사용자의 루틴 목록을 조회합니다.")
-    @GetMapping ("/all")
+    @Operation(summary = "모든 사용자의 루틴 목록 조회", description = "모든 사용자의 루틴 목록을 조회합니다.")
+    @GetMapping("/all")
     public ResponseEntity<?> getAllUsersRoutines() {
         List<RoutineDto> routines = workoutRoutineService.getAllUsersRoutines();
         if (routines.isEmpty()) {
@@ -222,10 +220,10 @@ public class MainController {
         return ResponseEntity.ok(summaryRoutines);
     }
 
-    @Operation (summary = "본인의 모든 루틴 목록 조회", description = "사용자의 모든 루틴 목록을 조회합니다.")
-    @GetMapping ("/routines/all")
+    @Operation(summary = "본인의 모든 루틴 목록 조회", description = "사용자의 모든 루틴 목록을 조회합니다.")
+    @GetMapping("/routines/all")
     public ResponseEntity<?> getAllRoutines(HttpServletRequest request) {
-        int userId = (Integer)request.getAttribute("userId");
+        int userId = (Integer) request.getAttribute("userId");
 
         List<RoutineDto> routines = workoutRoutineService.getAllRoutines(userId);
         if (routines.isEmpty()) {
@@ -237,10 +235,10 @@ public class MainController {
         return ResponseEntity.ok(summaryRoutines);
     }
 
-    @Operation (summary = "본인의 최근 5개 루틴 목록 조회", description = "사용자의 최근 5개 루틴 목록을 조회합니다.")
-    @GetMapping ("/routines")
+    @Operation(summary = "본인의 최근 5개 루틴 목록 조회", description = "사용자의 최근 5개 루틴 목록을 조회합니다.")
+    @GetMapping("/routines")
     public ResponseEntity<?> getAllRoutinesLimit(HttpServletRequest request) {
-        int userId = (Integer)request.getAttribute("userId");
+        int userId = (Integer) request.getAttribute("userId");
 
         List<RoutineDto> routines = workoutRoutineService.getAllRoutinesLimit(userId);
         if (routines.isEmpty()) {
